spilt models.py into two models file - shopcart model - item model
