+++ service/common/error_handlers.py
@@ -85,6 +85,21 @@ def mediatype_not_supported(error):
     )
 
 
+
+@app.errorhandler(status.HTTP_409_CONFLICT)
+def resource_conflict(error):
+    """Handles conflicts with 409_CONFLICT"""
+    message = str(error)
+    app.logger.warning(message)
+    return (
+        jsonify(
+            status=status.HTTP_409_CONFLICT,
+            error="Conflict",
+            message=message,
+        ),
+        status.HTTP_409_CONFLICT,
+    )
+
 @app.errorhandler(status.HTTP_500_INTERNAL_SERVER_ERROR)
 def internal_server_error(error):
     """Handles unexpected server error with 500_SERVER_ERROR"""
